requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2024.1
requests==2.32.3
beautifulsoup4==4.12.3
PyPDF2==3.0.1
